[SHIM[HACK][REVERT-ME]] Add shim to change SDK version for  Supercell games
The Supercell games don't load on our devices. This is from updates
of augustus to those games. If we change the sdk to 26 or lower,
the games do load again. What the specific error is, we don't know.

So tell every Supercell game that we have an SDK of 26. This should work
just fine for every device.